requirements.txt
fastapi
uvicorn
google-genai
playwright==1.40.0
requests
pandas
numpy
PyPDF2
beautifulsoup4
lxml
matplotlib
scikit-learn